Events building block for multiple applications
Multiple skin doses can be added using this building block.  

At the dose application time, any permeant in the surface pool or vehicle is removed, followed by an immediate addition of a user-specified new dose.

The timing of the new dose is entered in skin_events...Applications...skin_application...[NAME OF EVENT GROUP]...Dose_time

The user can enter the dose mass per unit area in two ways:
1) in the skin_spatial_structure building block, which is then automatically used by the event group found under skin_events...Applications...skin_application...dose_from_spatial_structure_at_0_hours...Mass_dose_per_area

or

2) manually, in the event group itself as in skin_events...Applications...skin_application...dose_added_manually_at_12_hours...Mass_dose_per_area
 
More applications can be added by saving one of the two event groups as a pkml and reloading it under skin_application.

Remember to set the molecule start values to zero as the dose is no longer being set there, but in this new Events building block.